Chem

The Chemistry module of Jems handles all things related to chemical mixtures and isotopes.

Jems.Chem.Isotope — Type

struct Isotope

Structure containing basic info of an isotope:

Z: atomic number (# protons)

A: mass number (# protons + neutrons)

name: its name (eg "Hydrogen")

mass: atomic weight in amu

Jems.Chem.get_isotope_list — Method

get_isotope_list()

Returns a dictionary of all included isotopes in Jems, mapping symbols to Isotope objects.