Constants

The Constants module of Jems defines a common set of physical and mathematical constants for use across the package. Currently not all constants are exported, so some of them need to be accesed by using the module name (e.g. Const.FINE).

Jems.Constants.AMU

Jems.Constants.AU

Jems.Constants.AVO

Jems.Constants.CGAS

Jems.Constants.CGRAV

Jems.Constants.CLIGHT

Jems.Constants.CRAD

Jems.Constants.DAYYEAR

Jems.Constants.EV_TO_ERG

Jems.Constants.FINE

Jems.Constants.G_TIMES_MEARTH

Jems.Constants.G_TIMES_MJUPITER

Jems.Constants.G_TIMES_MSUN

Jems.Constants.HBAR

Jems.Constants.K_BOLTZ

Jems.Constants.LSUN

Jems.Constants.LY

Jems.Constants.ME

Jems.Constants.MEARTH

Jems.Constants.MEV_TO_ERGS

Jems.Constants.MJUPITER

Jems.Constants.MN

Jems.Constants.MP

Jems.Constants.MSUN

Jems.Constants.PLANCK_H

Jems.Constants.QE

Jems.Constants.RBOHR

Jems.Constants.RSUN

Jems.Constants.SECYEAR

Jems.Constants.SIGMA_SB

Jems.Constants.AMU — Constant

AMU = 1.6605390666e-24

Atomic mass unit in units of $\mathrm{g}$

Source: CODATA 2018

Has a relative uncertainty of $3\times10^{-10}$

Jems.Constants.AU — Constant

AU = 1.49597870700e13

Astronomical unit in $\mathrm{cm}$.

From "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy".

Jems.Constants.AVO — Constant

AVO = 6.02214076e23

Avogadro's constant in units of $\mathrm{mol}^{-1}$.

Source CODATA 2018

Exact constant

Jems.Constants.CGAS — Constant

CGAS = K_BOLTZ * AVO = 8.31446261815324e7

Ideal gas constant in units of $\mathrm{erg}~\mathrm{K}^{-1}~\mathrm{mol}^{-1}$

Exact constant

Jems.Constants.CGRAV — Constant

CGRAV = 6.67430e-8

Gravitational constant in units of $\mathrm{g}^{-1}~\mathrm{cm}^{3}~\mathrm{s}^{-2}$

Source: CODATA 2018

Has a relative uncertainty of $2.2\times10^{-5}$

Jems.Constants.CLIGHT — Constant

CLIGHT = 2.99792458e10

Speed of light in units of $\mathrm{cm}~\mathrm{s}^{-1}$.

Source CODATA 2018

Exact constant

Jems.Constants.CRAD — Constant

CRAD = SIGMA_SB * 4 / CLIGHT = 7.565733250280006e-15

Radiation density constant in units of $\mathrm{erg}~\mathrm{cm}^{-3}\mathrm{K}^{-4}$

Jems.Constants.DAYYEAR — Constant

DAYYEAR = 365.25e0

Number of days in a year

Jems.Constants.EV_TO_ERG — Constant

EV_TO_ERG = 1.602176634e-12

Energy of $1~\mathrm{eV}$ in $\mathrm{erg}$

Source: CODATA 2018

Jems.Constants.FINE — Constant

FINE = QE^2 / (HBAR * CLIGHT) = 0.007297352565305213

Fine structure constant

Jems.Constants.G_TIMES_MEARTH — Constant

G_TIMES_MEARTH = 3.986004e20

Product of the gravitational constant times the mass of the Earth. In units of $\mathrm{cm}^3~\mathrm{s}^{-2}$.

Taken from the "IAU 2015 Resolution B3 on Recommended Nominal Conversion Constants for Selected Solar and Planetary Properties".

Jems.Constants.G_TIMES_MJUPITER — Constant

G_TIMES_MJUPITER = 1.2668653e23

Product of the gravitational constant times the mass of Jupiter. In units of $\mathrm{cm}^3~\mathrm{s}^{-2}$.

Taken from the "IAU 2015 Resolution B3 on Recommended Nominal Conversion Constants for Selected Solar and Planetary Properties".

Jems.Constants.G_TIMES_MSUN — Constant

G_TIMES_MSUN = 1.3271244e26

Product of the gravitational constant times the mass of the Sun. In units of $\mathrm{cm}^3~\mathrm{s}^{-2}$.

Taken from the "IAU 2015 Resolution B3 on Recommended Nominal Conversion Constants for Selected Solar and Planetary Properties".

Jems.Constants.HBAR — Constant

HBAR = PLANCK_H / (2 * pi) = 1.0545718176461565e-27

Reduced Planck constant in units of $\mathrm{erg}~\mathrm{s}$

Source: CODATA 2018

Exact constant

Jems.Constants.K_BOLTZ — Constant

K_BOLTZ = 1.380649e-16

Boltzmann constant in units of $\mathrm{erg}^{-1}~\mathrm{K}^{-1}$.

Source CODATA 2018

Exact constant

Jems.Constants.LSUN — Constant

LSUN = 3.828e33

Solar luminosity in $\mathrm{erg}~\mathrm{s}^{-1}$.

Taken from the "IAU 2015 Resolution B3 on Recommended Nominal Conversion Constants for Selected Solar and Planetary Properties".

Jems.Constants.LY — Constant

LY = CLIGHT*SECYEAR = 9.4607304725808e17

Lightyear in units of $\mathrm{cm}$

Jems.Constants.ME — Constant

ME = 9.1093837015e-28

Electron mass in units of $\mathrm{g}$

Source: CODATA 2018

Has a relative uncertainty of $3.0\times10^{-10}$

Jems.Constants.MEARTH — Constant

MEARTH = G_TIMES_MEARTH/CGRAV = 5.972167867791379e27

Mass of the Earth in $\mathrm{g}$ computed from the product GM.

Jems.Constants.MEV_TO_ERGS — Constant

MEV_TO_ERG = 1e6 * EV_TO_ERG = 1.602176634e-6

Energy of $1~\mathrm{MeV}$ in $\mathrm{erg}$

Jems.Constants.MJUPITER — Constant

MJUPITER = G_TIMES_MJUPITER/CGRAV = 1.89812459733605e30

Mass of Jupiter in $\mathrm{g}$ computed from the product GM.

Jems.Constants.MN — Constant

MN = 1.6749274980e-24

Neutron mass in units of $\mathrm{g}$

Source: CODATA 2018

Has a relative uncertainty of $5.7\times10^{-10}$

Jems.Constants.MP — Constant

MP = 1.67262192369e-24

Neutron mass in units of $\mathrm{g}$

Source: CODATA 2018

Has a relative uncertainty of $3.1\times10^{-10}$

Jems.Constants.MSUN — Constant

MSUN = G_TIMES_MSUN/CGRAV = 1.9884098706980504e33

Mass of the Sun in $\mathrm{g}$ computed from the product GM.

Jems.Constants.PLANCK_H — Constant

PLANCK_H = 6.62607015e-27

Planck constant in units of $\mathrm{erg}~\mathrm{s}$

Source: CODATA 2018

Exact constant

Jems.Constants.QE — Constant

QE = (CLIGHT/10)*1.602176634e-19 = 4.803204712570263e-10

Electron charge in esu==(g cm^3 s^-2)^(1/2) converted from the electron charge in Coulomb.

Source CODATA 2018

Exact constant

Jems.Constants.RBOHR — Constant

RBOHR = HBAR^2 / (ME * QE^2) = 5.291772111941798e-9

Bohr radius in units of $\mathrm{cm}$

Jems.Constants.RSUN — Constant

RSUN = 6.957e10

Solar radius in $\mathrm{cm}$.

Taken from the "IAU 2015 Resolution B3 on Recommended Nominal Conversion Constants for Selected Solar and Planetary Properties".

Jems.Constants.SECYEAR — Constant

SECYEAR = DAYYEAR * 24 * 3600 = 3.15576e7

Number of seconds in a year.

Jems.Constants.SIGMA_SB — Constant

SIGMA_SB = (2*π^5*K_BOLTZ^4)/(15*CLIGHT^2*PLANCK_H^3) = 5.67037441918443e-5

Stefan-Boltzmann constant in units of $\mathrm{erg}~\mathrm{cm}^{-2}\mathrm{K}^{-4}\mathrm{s}^{-1}$