Evolution

The Evolution module of Jems contains the basic tools needed to combine all other modules to perform stellar evolution. It allows a fully customizable definition of the equations that are solved together with their boundary conditions.

StellarModel.jl

Jems.Evolution.StellarModel — Type

mutable struct StellarModel{TN<:Real,TD<:Real,TEOS<:EOS.AbstractEOS,TKAP<:Opacity.AbstractOpacity}

An evolutionary model for a star, containing information about the star's current state, as well as the independent variables of the model and its equations.

The struct has four parametric types, TN for 'normal' numbers, TD for dual numbers used in automatic differentiation, TEOS for the type of EOS being used and TKAP for the type of opacity law being used.

Jems.Evolution.StellarModel — Method

StellarModel(varnames::Vector{Symbol}, structure_equations::Vector{Function},
             nvars::Int, nspecies::Int, nz::Int, eos::AbstractEOS, opacity::AbstractOpacity)

Constructor for a StellarModel instance, using varnames for the independent variables, functions of the structure_equations to be solved, number of independent variables nvars, number of species in the network nspecies number of zones in the model nz and an iterface to the EOS and Opacity laws.

Jems.Evolution.StellarStepInfo — Type

mutable struct StellarStepInfo{TN<:Real}

Information used for a simulation step. A single stellar model can have three different objects of type StellarStepInfo, containing information from the previous step, information right before the Newton solver, and information after the Newton solver has completed.

The struct has one parametric type, TN to represent 'normal' numbers. No fields here need to have dual numbers as these will not be used in automatic differentiation routines.

Jems.Evolution.TypeStableEquation — Type

struct TypeStableEquation{TS,TD<:Real}

Structure that wraps a stellar structure equation into a type stable object, using FunctionWrappers.jl. This requires that the stellar structure equations have the following signature:

function structure_equation(::TS, ::Int,
                            ::Matrix{TD}, ::Matrix{TD}, ::Matrix{TD},
                            ::EOSResults{TD}, ::EOSResults{TD}, ::EOSResults{TD}
                            ::TD, ::TD, ::TD)::TD

For typical usage, TS is the concrete type of StellarModel, and TD the type of dual number being used for automatic differentiation. The function must return an object of type TD, the result of the equation.

Jems.Evolution.init_diff_cache! — Method

init_diff_cache!(sm::StellarModel)

Initializes the diffcaches to the values of the independent variables, and sets ones in the correct spots where the dxi^k/dx_i^k entries lie.

Options.jl

Jems.Evolution.IOOptions — Type

mutable struct IOOptions

Substructure of Options containing controls relating to input/output of data

Jems.Evolution.Options — Type

mutable struct Options

Structure containing tweakable controls of Jems.

Jems.Evolution.SolverOptions — Type

mutable struct SolverOptions

Substructure of Options containing controls relating to the Newton solver

Jems.Evolution.TerminationOptions — Type

mutable struct TerminationOptions

Substructure of Options containing controls relating to termination of the simulation

Jems.Evolution.TimestepOptions — Type

mutable struct TimestepOptions

Substructure of Options containing controls relating to timestepping

Jems.Evolution.set_options! — Method

set_options!(opt::Options, toml_path::String)

Sets the controls in opt to the values supplied in the TOML file toml_path, containing key: value pairs. Invalid keys are not allowed, and an Exception will be thrown.

Equations.jl

Jems.Evolution.equationContinuity — Method

equationContinuity(sm::StellarModel, k::Int,
                   varm1::Matrix{TT}, var00::Matrix{TT}, varp1::Matrix{TT},
                   eosm1::EOSResults{TT}, eos00::EOSResults{TT}, eosp1::EOSResults{TT},
                   κm1::TT, κ00::TT, κp1::TT)::TT where {TT<:Real}

Default equation of mass continuity, evaluated for cell k of StellarModel sm.

Arguments

Identical to equationHSE for compatibility with TypeStableEquation

Returns

Residual of comparing dr^3/dm with 3/(4πρ)

Jems.Evolution.equationH1 — Method

equationH1(sm::StellarModel, k::Int,
           varm1::Matrix{TT}, var00::Matrix{TT}, varp1::Matrix{TT},
           eosm1::EOSResults{TT}, eos00::EOSResults{TT}, eosp1::EOSResults{TT},
           κm1::TT, κ00::TT, κp1::TT)::TT where {TT<:Real}

Default equation of Hydrogen 1 evoluation, evaluated for cell k of StellarModel sm.

Arguments

Identical to equationHSE for compatibility with TypeStableEquation

Returns

Residual of comparing dX_H1/dt with its computed reaction rate

Jems.Evolution.equationHSE — Method

equationHSE(sm::StellarModel, k::Int,
            varm1::Matrix{TT}, var00::Matrix{TT}, varp1::Matrix{TT},
            eosm1::EOSResults{TT}, eos00::EOSResults{TT}, eosp1::EOSResults{TT},
            κm1::TT, κ00::TT, κp1::TT)::TT where {TT<:Real}

Default equation of hydrostatic equilibrium. Evaluates for cell k of StellarModel sm to what degree hydrostatic equilibrium is satisfied.

Arguments

sm: Stellar Model

k: cell number to consider

varm1: Matrix holding the dual numbers of the previous cell (k-1)

var00: Matrix holding the dual numbers of this cell (k)

varp1: Matrix holding the dual numbers of the next cell (k+1)

eosm1: EOSResults object holding the results of the EOS evaluation of the previous cell (k-1)

eos00: EOSResults object holding the results of the EOS evaluation of the current cell (k)

eosp1: EOSResults object holding the results of the EOS evaluation of the next cell (k+1)

κm1: Opacity evaluated at the previous cell (k-1)

κ00: Opacity evaluated at the current cell (k)

κp1: Opacity evaluated at the next cell (k+1)

Returns

Residual of comparing dlnP/dm with -GM/4πr^4, where the latter is evaluated at the face of cell k and k+1.

Jems.Evolution.equationHe4 — Method

equationHe4(sm::StellarModel, k::Int,
           varm1::Matrix{TT}, var00::Matrix{TT}, varp1::Matrix{TT},
           eosm1::EOSResults{TT}, eos00::EOSResults{TT}, eosp1::EOSResults{TT},
           κm1::TT, κ00::TT, κp1::TT)::TT where {TT<:Real}

Default equation of Helium 4 evoluation, evaluated for cell k of StellarModel sm.

Arguments

Identical to equationHSE for compatibility with TypeStableEquation

Returns

Residual of comparing He4 content with 1 minus the H1 content.

Jems.Evolution.equationLuminosity — Method

equationLuminosity(sm::StellarModel, k::Int,
                   varm1::Matrix{TT}, var00::Matrix{TT}, varp1::Matrix{TT},
                   eosm1::EOSResults{TT}, eos00::EOSResults{TT}, eosp1::EOSResults{TT},
                   κm1::TT, κ00::TT, κp1::TT)::TT where {TT<:Real}

Default equation of energy generation, evaluated for cell k of StellarModel sm.

Arguments

Identical to equationHSE for compatibility with TypeStableEquation

Returns

Residual of comparing dL/dm with ϵnuc - cₚ * dT/dt - (δ / ρ) * dP/dt

Jems.Evolution.equationT — Method

equationT(sm::StellarModel, k::Int,
          varm1::Matrix{TT}, var00::Matrix{TT}, varp1::Matrix{TT},
          eosm1::EOSResults{TT}, eos00::EOSResults{TT}, eosp1::EOSResults{TT},
          κm1::TT, κ00::TT, κp1::TT)::TT where {TT<:Real}

Default equation of energy transport, evaluated for cell k of StellarModel sm.

Arguments

Identical to equationHSE for compatibility with TypeStableEquation

Returns

Residual of comparing dlnT/dm with -∇*GMT/4πr^4P, where the latter is evaluation at the face of cell k and k+1.

Solver.jl

EvolutionLoop.jl

Jems.Evolution.cycle_step_info! — Method

cycle_step_info!(sm::StellarModel)

Moves the model info of the StellarModel sm over one state: start step info -> end step info -> previous step info -> start step info.

Jems.Evolution.do_evolution_loop — Method

do_evolution_loop(sm::StellarModel)

Performs the main evolutionary loop of the input StellarModel sm. It continues taking steps until one of the termination criteria is reached (defined in sm.opt.termination).

Jems.Evolution.get_dt_next — Method

get_dt_next(sm::StellarModel)

Computes the timestep of the next evolutionary step to be taken by the StellarModel sm by considering all timestep controls (sm.opt.timestep).

Jems.Evolution.set_end_step_info! — Method

set_end_step_info(sm::StellarModel)

Sets the end step info (sm.esi) from current state of the StellarModel sm.

Jems.Evolution.set_start_step_info! — Method

set_start_step_info!(sm::StellarModel)

Sets the start step info of the StellarModel sm by copying from the previous step info.

InitialCondition.jl

Jems.Evolution.get_logdq — Method

get_logdq(k::Int, nz::Int, logdq_low::TT, logdq_high::TT, numregion::Int)::TT where {TT<:Real}

Computes the logarithm mass chunk logdq for zone k of a profile with total zones nz, while keeping in mind to better resolve the first and last numregion zones of the profile. It linearly interpolates the value from the inputs logdq_low and logdq_high in these regions, while keeping logdq_high in the middle zones.

Jems.Evolution.n1_polytrope_initial_condition — Method

n1_polytrope_initial_condition(sm::StellarModel, M::Real, R::Real; initial_dt=100 * SECYEAR)

Initializes a stellar model sm with values corresponding to an n=1 polytrope, setting the independent variables sm.ind_vars, etc. accordingly. Also sets the initial timestep to be taken, initial_dt.

Jems.Evolution.θ_n — Method

θ_n(ξ) = sin(ξ) / ξ, the sinc function

IO.jl

Jems.Evolution.get_history_dataframe_from_hdf5 — Method

get_history_dataframe_from_hdf5(hdf5_filename)

Returns a DataFrame object built from an hdf5 file, named hdf5_filename.

Jems.Evolution.get_profile_dataframe_from_hdf5 — Method

get_profile_dataframe_from_hdf5(hdf5_filename, profile_name)

Returns a DataFrame object built from an hdf5 file, named hdf5_filename, considering the column named profile_name

Jems.Evolution.get_profile_names_from_hdf5 — Method

get_profile_names_from_hdf5(hdf5_filename)

Retruns the column names of the profile data contained in the hdf5 file hdf5_filename.

Jems.Evolution.history_get_ind_vars_edge_value — Method

history_get_ind_vars_edge_value(sm::StellarModel, var_symbol::Symbol, edge::Symbol)

Returns the value of the independent variable var_symbol at either the surface or the center of the StellarModel sm. edge can be either :surface or :center.

Jems.Evolution.profile_get_ind_vars_value — Method

profile_get_ind_vars_value(sm::StellarModel, var_symbol::Symbol, k::Int)

Returns the value of the variable Symbol var_symbol at cell number k of the StellarModel sm.

Jems.Evolution.write_data — Method

write_data(dm::StellarModel)

Saves data (history/profile) for the current model, as required by the settings in sm.opt.io.