Plotting

This module provides live plotting routines for simulations

Plotting.jl

Jems.Plotting.end_of_evolution — Method

end_of_evolution(sm::StellarModel)

Perform end of evolution actions

Jems.Plotting.update_plotting! — Method

update_plots!(sm::StellarModel)

Updates all plots currently being displayed, by collecting appropriate data and notifying observables

Init.jl

Jems.Plotting.init_figure! — Method

init_figure!(sm::StellarModel)

Initializes the plotting figure and adds axes

Jems.Plotting.init_plots! — Method

init_plots(sm::StellarModel)

Sets up all observables to be traced this run, and creates the figure and axis where they will be plotted

History.jl

Jems.Plotting.create_history_observables! — Method

create_history_observables!(sm::StellarModel, plot::StellarModels.JemsPlot)

creates the x and y observables and adds them to the observable list of the given plot

Jems.Plotting.make_history_plot! — Method

init_HR_plot!(ax::Axis, xval::Observable, yval::Observable; line_kwargs=Dict())

Sets up the plot elements for an history plot

Jems.Plotting.update_history_plot! — Method

function update_history_plot!(plot::StellarModels.JemsPlot, sm::StellarModel)

Updates the given plot with relevant history data from the stellar model sm.

HRD.jl

Jems.Plotting.create_HR_observables! — Method

create_HR_observables!(sm::StellarModel, plot::StellarModels.JemsPlot)

Creates teff and L observables and adds them to the observable list of the given plot

Jems.Plotting.make_HR_plot! — Method

init_HR_plot!(ax::Axis, Teff::Observable, L::Observable, Teff_now::Observable, L_now::Observable;
                  line_kwargs=Dict(), scatter_kwargs=Dict())

Sets up the plot elements for an HRD

Jems.Plotting.update_HR_plot! — Method

function update_HR_plot!(plot::StellarModels.JemsPlot, sm::StellarModel)

Updates the given plot with the relevant HR data from the properties of the stellar model props.

Profile.jl

Jems.Plotting.create_profile_observables! — Method

function create_profile_observables!(plot::StellarModels.JemsPlot, xvals::Dict, yvals::Dict;
                                     alt_yvals::Dict=nothing)

Creates relevant profile observables for this plot given xvals, yvals and optionally alt_yvals for the right hand axis.

Jems.Plotting.make_profile_plot! — Method

function make_profile_plot!(ax::Axis, xvals::Observable, yvals::Dict{Symbol,Observable};
                            xlabel::AbstractString="", ylabels::Dict{Symbol,<:AbstractString}=Dict(),
                            alt_ax::Axis=nothing, alt_yvals::Dict{Symbol,Observable}=nothing,
                            alt_ylabels::Dict{Symbol,<:AbstractString}=nothing,
                            line_kwargs=Dict())

Plots a line for each entry in the yvals dict, and puts the given ylabels in a legend. The alt_ax is optional for the right hand side axis.

Jems.Plotting.update_profile_plot! — Method

update_profile_plot!(plot::StellarModels.JemsPlot, sm::StellarModel)

updates the observables of this plot with relevant data of the stellar model sm.