StellarModels

The StellarModels module of Jems defines models that it can evolve, along with their properties, initialization, IO operations, and option configuration.

StellarModel.jl

Jems.StellarModels.StellarModel — Type

mutable struct StellarModel{TN<:Real,TD<:Real,TEOS<:EOS.AbstractEOS,TKAP<:Opacity.AbstractOpacity}

An evolutionary model for a star, containing information about the star's current state, as well as the independent variables of the model and its equations.

The struct has four parametric types, TN for 'normal' numbers, TD for dual numbers used in automatic differentiation, TEOS for the type of EOS being used and TKAP for the type of opacity law being used.

Jems.StellarModels.StellarModel — Method

StellarModel(varnames::Vector{Symbol}, structure_equations::Vector{Function},
            nvars::Int, nspecies::Int, nz::Int, eos::AbstractEOS, opacity::AbstractOpacity)

Constructor for a StellarModel instance, using varnames for the independent variables, functions of the structure_equations to be solved, number of independent variables nvars, number of species in the network nspecies number of zones in the model nz and an iterface to the EOS and Opacity laws.

Jems.StellarModels.cycle_props! — Method

cycle_props!(sm::StellarModel)

Moves the model properties of the StellarModel sm over one state: startstepprops -> props -> prvstepprops -> startstepprops

Jems.StellarModels.uncycle_props! — Method

uncycle_props!(sm::StellarModel)

Moves the model properties of the StellarModel sm back one state: startstepprops <- props <- prvstepprops <- startstepprops

StellarModelProperties.jl

Jems.StellarModels.copy_mesh_properties! — Method

function copy_mesh!(sm, props_in, props_out)

Copies over the mesh quantities, i.e., nz, m, dm, and ind_vars from props_in into props_out.

Jems.StellarModels.evaluate_stellar_model_properties! — Method

function evaluate_stellar_model_properties!(sm, props::StellarModelProperties)

Evaluates the stellar model properties props from the ind_vars array. The goal is to save the 'state' of the StellarModel so we can easily get properties like rates, eos, opacity values, and retrace if a retry is called. This does not update the mesh/ind_vars arrays.

OneZone.jl

Jems.StellarModels.OneZone — Type

mutable struct OneZone{TNUMBER<:Real,TDUALFULL<:ForwardDiff.Dual,
                          TPROPS<:StellarModels.AbstractModelProperties,
                          TNET<:NuclearNetworks.AbstractNuclearNetwork,
                          TSOLVER<:StellarModels.AbstractSolverData}

Structure definition of a model having one internal zone.

Jems.StellarModels.OneZone — Type

OneZone(varnames::Vector{Symbol}, composition_equation::Function,
            nvars::Int, nspecies::Int)

Constructor for a OneZone instance, using varnames for the independent variables, the composition equation to be solved, number of independent variables nvars, number of species in the network nspecies

Options.jl

Jems.StellarModels.IOOptions — Type

mutable struct IOOptions

Substructure of Options containing controls relating to input/output of data

Jems.StellarModels.Options — Type

mutable struct Options

Structure containing tweakable controls of Jems.

Jems.StellarModels.PhysicsOptions — Type

mutable struct PhysicsOptions

Options that affect the physics of the computed model

Jems.StellarModels.RemeshOptions — Type

mutable struct RemeshOptions

Substructure of Options containing controls relating to remeshing

Jems.StellarModels.SolverOptions — Type

mutable struct SolverOptions

Substructure of Options containing controls relating to the Newton solver

Jems.StellarModels.TerminationOptions — Type

mutable struct TerminationOptions

Substructure of Options containing controls relating to termination of the simulation

Jems.StellarModels.TimestepOptions — Type

mutable struct TimestepOptions

Substructure of Options containing controls relating to timestepping

Jems.StellarModels.set_options! — Method

set_options!(opt::Options, toml_path::String)

Sets the controls in opt to the values supplied in the TOML file toml_path, containing key: value pairs. Invalid keys are not allowed, and an Exception will be thrown.

InitialCondition.jl

IO.jl

Jems.StellarModels.create_output_files! — Method

create_output_files(sm::StellarModel)

Creates output files for history and profile data

Jems.StellarModels.get_history_dataframe_from_hdf5 — Method

get_history_dataframe_from_hdf5(hdf5_filename)

Returns a DataFrame object built from an hdf5 file, named hdf5_filename.

Jems.StellarModels.get_ith_partial_history_dataframe_from_hdf5 — Method

get_ith_partial_history_dataframe_from_hdf5(hdf5_filename, i)

Returns a DataFrame object built from the ith partial of an hdf5 file, named hdf5_filename.

Jems.StellarModels.get_ith_partial_profile_names_from_hdf5 — Method

get_ith_partial_profile_names_from_hdf5(hdf5_filename, i)

Returns the column names of the ith partial of profile data contained in the hdf5 file hdf5_filename.

Jems.StellarModels.get_profile_dataframe_from_hdf5 — Method

get_profile_dataframe_from_hdf5(hdf5_filename, profile_name)

Returns a DataFrame object built from an hdf5 file, named hdf5_filename, considering the column named profile_name

Jems.StellarModels.get_profile_names_from_hdf5 — Method

get_profile_names_from_hdf5(hdf5_filename)

Returns the column names of the profile data contained in the hdf5 file hdf5_filename.

Jems.StellarModels.write_data — Method

write_data(sm::StellarModel)

Saves data (history/profile) for the current model, as required by the settings in sm.opt.io.

Remesher.jl

Jems.StellarModels.adjust_props_size! — Method

adjust_props_size(sm, new_nz::Int, nextra::Int)

Returns a new StellarModelProperties object with an adjusted size. Copies over the following from the currect active properties:

nz

dt

time

ind_vars

mstar

m

dm

Jems.StellarModels.remesher! — Method

remesher!

Acts on sm.startstepprops using info from sm.prvstepprops to decide whether to merge/split cells

PlotterInterface.jl