Turbulence

The Turbulence module of Jems handles all things related to turbulence and convection inside stars.

Jems.Turbulence.AbstractTurb — Type

abstract type AbstractTurb

Abstract supertype from which all turbulence models must derive, ie:

struct MyTurb <: AbstractTurb

Jems.Turbulence.BasicMLT — Type

struct AdiabaticMLT <: AbstractTurb

Definitions used for MLT without inclusion of radiative losses

Jems.Turbulence.TurbResults — Type

mutable struct TurbResults{T1}

Structure that holds various results from the evaluation of the turbulence model of a certain cell.