EOS

The EOS module of Jems handles all things related to the equation of state of stellar matter.

Jems.EOS.AbstractEOS — Type

abstract type AbstractEOS

Abstract supertype from which all equations of state definitions must derive, ie:

struct MyEOS <: AbstractEOS

Jems.EOS.EOSResults — Type

mutable struct EOSResults{T1<:Real}

Structure that holds various results from the evaluation of the EOS of a certain cell.

Jems.EOS.IdealEOS — Type

struct IdealEOS <: AbstractEOS

Interface of an Ideal gas equation of state

Jems.EOS.get_μ_IdealEOS — Method

get_μ_IdealEOS(xa, species)

computes the molecular weight of the mixture xa, given and list of species.

Jems.EOS.set_EOS_resultsTρ! — Method

set_EOS_resultsTP!(eos::IdealEOS, r::EOSResults{TT}, lnT::TT, lnP::TT, xa::AbstractVector{TT},
                       species::Vector{Symbol}) where {TT<:Real}

Computes thermodynamical quantities of a mixture xa at temperature lnT and pressure lnP, given the ideal equation of state eos and list of species. The results are stored in the EOSResults object r.