EOS

The EOS module of Jems handles all things related to the equation of state of stellar matter.

Jems.EOS.IdealEOS — Type

struct IdealEOS <: AbstractEOS

Interface of an Ideal gas equation of state

Jems.EOS.get_μ_IdealEOS — Method

get_μ_IdealEOS(xa, species)

computes the molecular weight of the mixture xa, given and list of species.

Jems.EOS.set_EOS_resultsTP! — Method

get_EOS_resultsTP(eos, lnT, lnP, xa, species)

computes thermodynamical quantities of a mixture xa at temperature lnT and pressure lnP, given the ideal equation of state eos, list of species.