Evolution

The Evolution module of Jems contains the basic tools needed to combine all other modules to perform stellar evolution. It allows a fully customizable definition of the equations that are solved together with their boundary conditions.

StellarModel.jl

Jems.Evolution.StellarModel — Type

mutable struct StellarModel{T1<:Real, T2<: Real}

An evolutionary model for a star, containing information about the star's current state, as well as the independent variables of the model and its equations.

The struct has two parametric types, T1 for 'normal' numbers, T2 for dual numbers used in automatic differentiation

Jems.Evolution.StellarModel — Method

StellarModel(varnames::Vector{Symbol}, structure_equations::Vector{Function},
    nvars::Int, nspecies::Int, nz::Int, eos::AbstractEOS, opacity::AbstractOpacity)

Constructor for a StellarModel instance, using varnames for the independent variables, functions of the structure_equations to be solved, number of independent variables nvars, number of species in the network nspecies number of zones in the model nz and an iterface to the EOS and Opacity laws.

Jems.Evolution.StellarStepInfo — Type

mutable struct StellarStepInfo

Information used for a simulation step. A single stellar model can have three different objects of type StellarStepInfo, containing information from the previous step, information right before the Newton solver, and information after the Newton solver has completed.

The struct has one parametric type, T1 to represent 'normal' numbers. No fields here need to have dual numbers as these will not be used in automatic differentiation routines.

Jems.Evolution.init_diff_cache! — Method

init_diff_cache!(sm::StellarModel, k::Int)

Initializes the diffcaches to the values of the independent variables, and sets 1s in the correct spots where the dxi^k/dx_i^k entries lie.

Options.jl

Equations.jl

Solver.jl

EvolutionLoop.jl

Jems.Evolution.cycle_step_info! — Method

cycle_step_info!(sm::StellarModel)

Moves the model info of the StellarModel sm over one state: start step info -> end step info -> previous step info -> start step info

Jems.Evolution.do_evolution_loop — Method

do_evolution_loop(sm::StellarModel)

Performs the main evolutionary loop of the input StellarModel sm. It continues taking steps until one of the termination criteria is reached (defined in sm.opt.termination)

Jems.Evolution.get_dt_next — Method

get_dt_next(sm::StellarModel)

Computes the timestep of the next evolutionary step to be taken by the StellarModel sm.

Jems.Evolution.set_end_step_info! — Method

set_end_step_info(sm::StellarModel)

Sets the end step info (sm.esi) from current state of the StellarModel sm.

Jems.Evolution.set_start_step_info! — Method

set_start_step_info!(sm::StellarModel)

Sets the start step info of the StellarModel sm.

InitialCondition.jl

IO.jl

Jems.Evolution.history_get_ind_vars_edge_value — Method

history_get_ind_vars_edge_value(sm::StellarModel, var_symbol::Symbol, edge::Symbol)

Returns the value of the independent variable var_symbol at either the surface or the center of the StellarModel sm. edge can be either :surface or :center.

Jems.Evolution.profile_get_ind_vars_value — Method

profile_get_ind_vars_value(sm::StellarModel, var_symbol::Symbol, k::Int)

Returns the value of the variable Symbol var_symbol at cell number k of the StellarModel sm.