StellarEvolution

The StellarEvolution module contains the basic tools needed to combine all other modules to perform stellar evolution. It allows a fully customizable definition of the equations that are solved together with their boundary conditions.

StellarEvolution

StellarModel.jl

Options.jl

Equations.jl

Solver.jl

EvolutionLoop.jl

InitialCondition.jl

IO.jl

A model is initialized by specifying its independent variables and equations. The following creates a model with a basic ideal gas EOS and electron scattering opacity, which is then initialized using an n=1 polytrope of 1$M_\odot$ and $100R_\odot$. Options for a simulation can be specified with a toml file. Below we directly create a file with custom options and then load it up into the model. After setting everything up the simulation is run.

using Jems.Evolution
using Jems.EOS
using Jems.Opacity
using Jems.Constants

nvars = 6
nspecies = 2
varnames = [:lnP, :lnT, :lnr, :lum, :H1, :He4]
structure_equations=[Evolution.equationHSE, Evolution.equationT,
                        Evolution.equationContinuity, Evolution.equationLuminosity,
                        Evolution.equationH1, Evolution.equationHe4]
nz = 1000
eos = EOS.IdealEOS(false)
opacity = Opacity.SimpleElectronScatteringOpacity()
sm = StellarModel(varnames, structure_equations, nvars, nspecies, nz, eos, opacity)

#Initialize the model as n=1 polytrope with an initial timestep of 10 years
Evolution.n1_polytrope_initial_condition(sm, MSUN, 100*RSUN; initial_dt=10*SECYEAR)

#Load custom options
open("options.toml","w") do file
    write(file,"""
            [termination]
            max_model_number = 300
""")
end
Evolution.set_options!(sm.opt, "options.toml")

#run simulation
#Evolution.do_evolution_loop(sm)

The results of the simulation are provided in HDF5 format.

StellarModel.jl

Jems.Evolution.StellarModel — Type

mutable struct StellarModel

An evolutionary model for a star, containing information about the star's current state, as well as the independent variables of the model and its equations.

Jems.Evolution.StellarModel — Method

StellarModel(varnames::Vector{Symbol}, structure_equations::Vector{Function}, nvars::Int, 
nspecies::Int, nz::Int, eos::AbstractEOS, opacity::AbstractOpacity)

Constructor for a StellarModel instance, using varnames for the independent variables, functions of the structure_equations to be solved, number of independent variables nvars, number of species in the network nspecies number of zones in the model nz and an iterface to the EOS and Opacity laws.

Jems.Evolution.StellarStepInfo — Type

mutable struct StellarStepInfo

Information used for a simulation step. A single stellar model can have three different objects of type StellarStepInfo, containing information from the previous step, information right before the Newton solver, and information after the Newton solver has completed.

Options.jl

Equations.jl

Solver.jl

Jems.Evolution.eval_cell_eqs — Method

eval_cell_eqs(sm::StellarModel, k::Int, ind_vars_view::Vector{<:TT}) where{TT<:Real}

Evaluates the stellar structure equations of the stellar model, sm, at cell k, given the view of the independent variables, ind_vars_view.

Jems.Evolution.eval_eqs! — Method

eval_eqs(sm::StellarModel)

Evaluates the stellar structure equations of the stellar model, sm, for all cells.

Jems.Evolution.eval_jacobian! — Method

eval_jacobian!(sm::StellarModel)

Evaluates the whole Jacobian matrix of the given StellarModel sm.

Jems.Evolution.eval_jacobian_row! — Method

eval_jacobian_row!(sm::StellarModel, k::int)

Evaluates row k of the Jacobian matrix of the given StellarModel sm.

EvolutionLoop.jl

Jems.Evolution.cycle_step_info! — Method

cycle_step_info!(sm::StellarModel)

Moves the model info of the StellarModel sm over one state: start step info -> end step info -> previous step info -> start step info

Jems.Evolution.do_evolution_loop — Method

do_evolution_loop(sm::StellarModel)

Performs the main evolutionary loop of the input StellarModel sm. It continues taking steps until one of the termination criteria is reached (defined in sm.opt.termination)

Jems.Evolution.get_dt_next — Method

get_dt_next(sm::StellarModel)

Computes the timestep of the next evolutionary step to be taken by the StellarModel sm.

Jems.Evolution.set_end_step_info! — Method

set_end_step_info(sm::StellarModel)

Sets the end step info (sm.esi) from current state of the StellarModel sm.

Jems.Evolution.set_start_step_info! — Method

set_start_step_info!(sm::StellarModel)

Sets the start step info of the StellarModel sm.

InitialCondition.jl

IO.jl

Jems.Evolution.history_get_ind_vars_edge_value — Method

history_get_ind_vars_edge_value(sm::StellarModel, var_symbol::Symbol, edge::Symbol)

Returns the value of the independent variable var_symbol at either the surface or the center of the StellarModel sm. edge can be either :surface or :center.

Jems.Evolution.profile_get_ind_vars_value — Method

profile_get_ind_vars_value(sm::StellarModel, var_symbol::Symbol, k::Int)

Returns the value of the variable Symbol var_symbol at cell number k of the StellarModel sm.