Opacity

The Opacity module of Jems handles all things related to the opacity of stellar matter.

Jems.Opacity.AbstractOpacity — Type

abstract type AbstractOpacity

Abstract supertype from which all defined opacity laws must derive, ie:

struct MyOpacity <: AbstractOpacity

Jems.Opacity.get_opacity_resultsTρ — Method

get_opacity_resultsTP(opacity::SimpleElectronScatteringOpacity, lnT::TT, lnP::TT, xa::Vector{<:TT},
                        species::Vector{Symbol})::TT where {TT<:Real}

Evaluates the opacity of the current mixture with mass fractions xa, species symbols species (both these should be of length nspecies), the natural log of temperature and pressure lnT, lnP, and the opacity law opacity.