StellarModels

The Chemistry module of Jems handles all things related to chemical mixtures and isotopes.

StellarModel.jl

Jems.StellarModels.StellarModel — Type

mutable struct StellarModel{TN<:Real,TD<:Real,TEOS<:EOS.AbstractEOS,TKAP<:Opacity.AbstractOpacity}

An evolutionary model for a star, containing information about the star's current state, as well as the independent variables of the model and its equations.

The struct has four parametric types, TN for 'normal' numbers, TD for dual numbers used in automatic differentiation, TEOS for the type of EOS being used and TKAP for the type of opacity law being used.

Jems.StellarModels.StellarModel — Method

StellarModel(varnames::Vector{Symbol}, structure_equations::Vector{Function},
             nvars::Int, nspecies::Int, nz::Int, eos::AbstractEOS, opacity::AbstractOpacity)

Constructor for a StellarModel instance, using varnames for the independent variables, functions of the structure_equations to be solved, number of independent variables nvars, number of species in the network nspecies number of zones in the model nz and an iterface to the EOS and Opacity laws.

Jems.StellarModels.StellarStepInfo — Type

mutable struct StellarStepInfo{TN<:Real}

Information used for a simulation step. A single stellar model can have three different objects of type StellarStepInfo, containing information from the previous step, information right before the Newton solver, and information after the Newton solver has completed.

The struct has one parametric type, TN to represent 'normal' numbers. No fields here need to have dual numbers as these will not be used in automatic differentiation routines.

Jems.StellarModels.TypeStableEquation — Type

struct TypeStableEquation{TS,TD<:Real}

Structure that wraps a stellar structure equation into a type stable object, using FunctionWrappers.jl. This requires that the stellar structure equations have the following signature:

function structure_equation(::TS, ::Int,
                            ::Matrix{TD}, ::Matrix{TD}, ::Matrix{TD},
                            ::EOSResults{TD}, ::EOSResults{TD}, ::EOSResults{TD}
                            ::Matrix{TD},
                            ::TD, ::TD, ::TD)::TD

For typical usage, TS is the concrete type of StellarModel, and TD the type of dual number being used for automatic differentiation. The function must return an object of type TD, the result of the equation.

Jems.StellarModels.init_diff_cache! — Method

init_diff_cache!(sm::StellarModel)

Initializes the diffcaches to the values of the independent variables, and sets ones in the correct spots where the dxi^k/dx_i^k entries lie.

Options.jl

Jems.StellarModels.IOOptions — Type

mutable struct IOOptions

Substructure of Options containing controls relating to input/output of data

Jems.StellarModels.Options — Type

mutable struct Options

Structure containing tweakable controls of Jems.

Jems.StellarModels.RemeshOptions — Type

mutable struct RemeshOptions

Substructure of Options containing controls relating to remeshing

Jems.StellarModels.SolverOptions — Type

mutable struct SolverOptions

Substructure of Options containing controls relating to the Newton solver

Jems.StellarModels.TerminationOptions — Type

mutable struct TerminationOptions

Substructure of Options containing controls relating to termination of the simulation

Jems.StellarModels.TimestepOptions — Type

mutable struct TimestepOptions

Substructure of Options containing controls relating to timestepping

Jems.StellarModels.set_options! — Method

set_options!(opt::Options, toml_path::String)

Sets the controls in opt to the values supplied in the TOML file toml_path, containing key: value pairs. Invalid keys are not allowed, and an Exception will be thrown.

InitialCondition.jl

Jems.StellarModels.get_logdq — Method

get_logdq(k::Int, nz::Int, logdq_low::TT, logdq_high::TT, numregion::Int)::TT where {TT<:Real}

Computes the logarithm mass chunk logdq for zone k of a profile with total zones nz, while keeping in mind to better resolve the first and last numregion zones of the profile. It linearly interpolates the value from the inputs logdq_low and logdq_high in these regions, while keeping logdq_high in the middle zones.

Jems.StellarModels.n1_polytrope_initial_condition! — Method

n1_polytrope_initial_condition(sm::StellarModel, M::Real, R::Real; initial_dt=100 * SECYEAR)

Initializes a stellar model sm with values corresponding to an n=1 polytrope, setting the independent variables sm.ind_vars, etc. accordingly. Also sets the initial timestep to be taken, initial_dt.

Jems.StellarModels.θ_n — Method

θ_n(ξ) = sin(ξ) / ξ, the sinc function

IO.jl

Jems.StellarModels.get_history_dataframe_from_hdf5 — Method

get_history_dataframe_from_hdf5(hdf5_filename)

Returns a DataFrame object built from an hdf5 file, named hdf5_filename.

Jems.StellarModels.get_profile_dataframe_from_hdf5 — Method

get_profile_dataframe_from_hdf5(hdf5_filename, profile_name)

Returns a DataFrame object built from an hdf5 file, named hdf5_filename, considering the column named profile_name

Jems.StellarModels.get_profile_names_from_hdf5 — Method

get_profile_names_from_hdf5(hdf5_filename)

Retruns the column names of the profile data contained in the hdf5 file hdf5_filename.

Jems.StellarModels.history_get_ind_vars_edge_value — Method

history_get_ind_vars_edge_value(sm::StellarModel, var_symbol::Symbol, edge::Symbol)

Returns the value of the independent variable var_symbol at either the surface or the center of the StellarModel sm. edge can be either :surface or :center.

Jems.StellarModels.profile_get_ind_vars_value — Method

profile_get_ind_vars_value(sm::StellarModel, var_symbol::Symbol, k::Int)

Returns the value of the variable Symbol var_symbol at cell number k of the StellarModel sm.

Jems.StellarModels.write_data — Method

write_data(dm::StellarModel)

Saves data (history/profile) for the current model, as required by the settings in sm.opt.io.

Remesher.jl

Jems.StellarModels.remesher! — Method

remesher!

Todo